ack-tracker: fix LogSource lifetime in BatchedAckTracker
LogSource should be kept alive until we have
 * pending events (unfortunately when an event is posted, it will be
 called even after the event is unregistered)
 * running timers

Before posting the iv_ev_event, we should increase the refcount for
LogSource to guarantee that the LogSource is alive when the handler is
called.

Other part of the solution is to make sure that we have only one pending batch timer
restart requested event - otherwise we would increase the refcounter for
each full batches. This would be a problem in case of multiple pending
events, as the LogSource instance is kept alive only by these pending events and
the last timer restart event handler would free up the LogSource which means that
it would be a dangling pointer when the registered timer callback is called.

originally we had this in batch timer restart requested handler:
    if (!log_pipe_unref(&self->super.source->super))  ---> it won't true, due to other pending events
      _restart_batch_timer(self);

So if we have N batches that are full before the pending event is
executed, we would restart the timer N-1 times and the timer's cookie
would be a dangling pointer.

Signed-off-by: Laszlo Budai <laszlo.budai@outlook.com>