ack-tracker: BatchedAckTracker: ack tracker is destroyed on its own thread
AckTracker::manage_msg_ack() is called from destination threads and when
LogSource refcnt reach 0, it's free() method will destroy its
AckTracker.
The problem is that the pending_request_restart_timer_lock is locked by
another thread - and we cannot call g_mutex_clear() in this case.
As this mutex can only be locked by the timer at ack_tracker_free()
time, we have to move the destruction of the AckTacker to the thread
that started the LogSource.

Signed-off-by: Laszlo Budai <laszlo.budai@outlook.com>