python-source: add finalize() method
In some cases we have a client and ACKs are propagated back
to the message originator. In these cases clients cannot be closed
in a deinit() method as there can be pending ACKs even after deinit().
The finalize() method is for closing these clients when the
PythonSource is freed up.

Signed-off-by: Laszlo Budai <laszlo.budai@outlook.com>