configure: fix journald namespaces detection
With older pkg-config versions (0.27.1 on centos-7)
PKG_CONFIG_CHECK_MODULES() macro doesn't check the library's version
on multiple calls (if they have the same prefix).
It seems to be that the macro is idempotent.

With version 0.29.1 it is allowed to call the macro multiple times with the same prefix.

Alternatively, this could be solved without additional PKG_CHECK_MODULES() calls on libsystemd,
with calling $PKG_CONFIG --modversion directly if we detected libsystemd in the first place.

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>