nvtable: nv_table_ref only when compaction is needed
nv_table_ref was introduced in
https://github.com/syslog-ng/syslog-ng/commit/fc1c857e4d31bea4519059da94eb61359e51ff41
for a neat trick to avoid condition around nv_table_unref:
if the original nv_table_was reffed, then nv_table_unref could be
called unconditionally.

As ref_cnt is not atomic, this causes trouble in logstore, where if
multiple logstores are added into the same logpath, they will
serialialize the same message in parallel (if flags(serialized) is
used).

So instead, we add a different function to avoid the ref of the payload.

Signed-off-by: Antal Nemes <antal.nemes.hu@gmail.com>