light: remove SourceWriter
SourceWriter seems like an unnecessary layer.
It is not general enough, as it forces to have `path` in write_log.

SourceDrivers can simply use the necessary IO layer without the
limitations of SourceWriter.

DestinationReader should be removed as well, first, because of symmetry,
second, the `path` argument and `wait_for_creation()` shows that it only
works with FileIO, but it is a bigger refactor, which I did not want to
include in this patchset.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>