Mk/lex-rules.am: insert bison version number into the generated source
We need conditional compilation based on the bison version as some of the
internals have changed between 3.5.1 and 3.6.1 and we would like to support
at the very least 3.5.1 (as ubuntu focal uses that).

Instead of using the host's bison version use the one that we used to
generate the grammars, as this will be the case with generated tarballs with
a recent bison not available on the compiling host.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>