dbld/rules: allow the cached tarball to be outside of the build directory
In case we are attempting to build from a release artifact, the tarball
would be outside of the main syslog-ng directory, as we would extract that
to get the syslog-ng directory in the first place.

This patch adds a feature to dbld/rules namely that it will find this
release artifact and copy it to the right place to avoid having
to regenerate the tarball during the building of binary packages.

This is pretty useful in our github action based package building job,
but will also be useful for anyone trying to use release tarballs to
build binaries.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>