dbld/rules: don't mess with packaging files if they are already present
This patch adds a check to 'pkg-tarball' to see if packaging files are
already present. The aim here is to avoid having to fetch the
"tarball" image just to add the Debian and RPM packaging files. This
is not that much of a problem when using interactively on a dev workstation,
but shaves off 4-5 minutes in our CI.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>