logmsg: fix stringop-overflow warning
Latest versions of gcc now has warnings in place to detect string array
overflows (-Wstringop-overflow, defaults to enabled). This is
now triggered for zero-sized arrays, which we have used to to indicate
variable data "hanging off" at the end of structs, namely NVEntry.

This patch removes the zero size and starts using the C99 standard for
the same thing, this should work at the very least since gcc 4.5, but
quite possibly earlier than that.

https://gcc.gnu.org/c99status.html

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>