afuser: fix reading potentially uninitialized bytes after utmp->ut_line
ut->ut_line is not guaranteed to be zero-terminated and the size of the
character array is 32 bytes on Linux, which we intend to stuff into
an array [123, 128] in size. This might cause us to read over our
source buffer.

This is not an issue in practice as the array is zero padded and I am
set to see the name of a tty device file that is 32 bytes or more.

The issue was detected by a new gcc 10 warning -Wstringop-overflow.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>