doc: separate manpage installation from the building of manpages
In case of a release tarball, we already have the manpages, even without
documentation tools, so add a separate --enable-manpage-install option
that enables only the installation of manual pages.

This option automatically detects if we have the generated manpages or
we have the capability of building the manpages (e.g. enable-manpages is
yes)

This way the following use-cases are supported:
  1) fresh checkout, we don't care about manpages, neither building nor
     installing them, e.g. CI. (docbook is not needed)
  2) fresh checkout, we build & install manual pages (docbook is required)
  3) release tarball, we install manpages as shipped but without the
     ability to rebuild them (docbook is not needed)
  4) release tarball, we install manpages as shipped or rebuild them as
     needed (docbook is required)

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>