autotools: fix unit test wrapper script shadowing failed unit tests
The LOG_COMPILER wrapper script used to workaround a criterion bug,
which prevents to detect if the test case exit status shows an error.
This was introduced when autotools sanitizer support added.

However, the test-grep.sh wrapper script shadows the exit status of the test case,
so we need to propagate this back to Make.

autotools LOG_COMPILER documentation for reference:
https://www.gnu.org/software/automake/manual/html_node/Parallel-Test-Harness.html#Parallel-Test-Harness

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>