LightFile: move file.py from driver_io to common
It is not a DriverIO level class. The FileIO correctly uses the File
class, but DriverIO classes must only be used by syslog-ng Drivers.

Currently there are other classes, which use the FileIO, but they
should use File instead.

This commit is just the first step of this transition. It does not
change the user's of FileIO.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>
Signed-off-by: Andras Mitzki <andras.mitzki@balabit.com>