dbparser: remove timer_process_params from PatternDB
This is a refactor commit.

timer_process_params was a PatternDB field that was shared among threads,
but contained stack addresses (thead-local data), which can be dangerous,
because it requires proper locking and lifetime considerations.

This commit removes the field and uses the new parameter of
timer_wheel_set_time(): caller_context.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>