affile/wildcard: fix crash when reader replaced
The number of active readers are limited in wildcard file source, the
resot of the files that needs prceossing as stored in a waiting_list.

It is possible to have a reader for a file, and that same file path to
be in the waiting_list, in that case when (for example a reader space is
avaible, and new reader can be created) a new reader is created for the
file. Due to using a hashtable for storing the readers using the path as
key, the newly created reader and the old one wants to occupy the same
space.

The `g_hash_table_insert` (that is used here) is implemented as it
replaces the old one (and calls {key,value}_destroy_func) on the old
one.

In the implementation that meant to unref the reader, while this should
be okay due to reference counting. There is an extra requirement in the
order of freeing between FileReader and LogReader, namely LogReader has
a reference to FileReader, and LogReader should not be the one to free
(via unref or otherwise) the FileReader, as the later still has a
reference to it.

reproduction of the crash:

config:
```
@version: 7.0

log {

source {
  wildcard_file(
    base-dir("/tmp/input/")
    filename-pattern("*")
  );
};

destination {
  file("/tmp/output.txt");
};

flags(flow-control);

};
```
create a non empty file in /tmp/input/

1. start syslog-ng with the above configuration
2. start an infinite loop with renaming the file created in
   /tmp/input/
   (for example while true ; mv /tmp/input/2 /tmp/input1 && mv
/tmp/input/1 /tmp/input2 ; end)
The crash might occure right away, or after a few minutes.

Signed-off-by: Kokan <kokaipeter@gmail.com>