affile/file_list: iterator: begin,end,next
The PendingFileList only had interface for poping entries, but not
always the topmost entry is needed from a list. In that case the user
must pop and re-add entries later.

This new interfaces gives a choice to iterate through the entries, and
remove only the desired item.

A fix following this requires such change.

Signed-off-by: Kokan <kokaipeter@gmail.com>