msg-format: move LP_NO_MULTI_LINE processing to msg-format
This is a common option that can apply to any msg-format plugins, so
move it to msg-format.

Also, newline removal was not performed in case of flags(no-parse) or in
case of an error, which it should, as no-multi-line generally means
that the user does not want multi-line messages from a source at all.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>