logsource: add initial_window_size
initial_window_size stores the initial (static) window size of the host.

This is different than LogSourceOptions::init_window_size, because the
currently it is not supported to reconfigure either log-iw-size() or
the dynamic-window-size() option in case a source is kept alive.

initial_window_size represents the value that is actually in use.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>