logsource: fix crash with dynamic-window-size() when window is reconfigured
Currently, it is not supported to reconfigure either log-iw-size() or
the dynamic-window-size() option on a kept-alive source. The current
behavior is to maintain original window sizes when keep-alive() is enabled
for a network source. New connections and connections configured with
keep-alive(no) will, of course, use the recently configured window values.

This commit fixes a crash (underflow) in the dynamic window mechanism by
using the static window size that is actually in-use, not
LogSourceOptions::init_window_size, which may contain a different, recently
configured value.

Reproduction steps:
  log {
    source {
      network(port(4444) log-iw-size(4000) max-connections(10) dynamic-window-size(100000) keep-alive(yes));
    };
  };

1. Start syslog-ng with the config above
2. Connect to the source (nc localhost 4444)
3. Change log-iw-size() to a smaller value (log-iw-size(3000)) and reload
4. Close the connection that was made in step 2

Signed-off-by: László Várady <laszlo.varady@protonmail.com>