apphook: add AH_CONFIG_STOPPED hook
The new hook can be used to perform a callback when the old configuration
has stopped (ie threads exited) before being restarted.

This also moves the res_init() function to reinitialize the resolver,
as my suspicion is that this might equally trigger race conditions
to the timezone related globals (ie: tzname, see the subsequent commits).

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>