python: do not double free seqnum and template_options
When these two objects are passed to `PyObject_SetAttrString()`, their
reference counter is incremented. In the next line, we drop our own
reference to them with `Py_DECREF()`, but still keep track of them
to free it later in `_py_free_bindings()`.

It is not needed, the python lib will take care of the reference, that
was added in `PyObject_SetAttrString().

Because we do not need these objects in our code, it is the easiest,
to not store them in our struct.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>