persist-state: change PERSIST_FILE_WATERMARK size
Persist file uses a "pre-allocation" mechanism to circumvent some
side effects of a running out of space scenario.

More on that in PR #1785, especially in 45a4157.

Thus PERSIST_FILE_WATERMARK must be bigger than any entry we want to
store in the persist file. (Including the extra headers and flags.)

One of our internal module has an 8k buffer. Considering the extra
space for the headers, the current value is: 8k+256 bytes.

Signed-off-by: Laszlo Szemere <laszlo.szemere@oneidentity.com>