kafka-c: fix double ack on re-added messages
Delivery_report_callback re-adds the message to the head of the queue for resending.
(with more workers, the message is possibly re-added to a different worker thread's queue
than the original worker who tried to sent out the message,
as we always poll from worker#0 and re-add the message to worker#0's queue).

The problem was, that the message was re-added to the queue with acknowledgement re-enabled.
This can cause trouble in both flow-controlled and non flow-controlled cases, as
even with a flow-controlled log path, the message is already acknowledged since kafka-c driver's insert method returned with SUCCESS,
and delivery failure only detected in the delivery_report callback.

Acknowledgement in kafka-c has to be revised, so we can acknowledge at the right time.
This will be done in a later step.

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>