afuser: support time-reopen()
`afuser` always used a default 600 value to disconnect from the tty.
It is better to make it configurable. `time_reopen` is an option
which has a similar behavior in other drivers, so it is logical to
make this new option `time-reopen()`.

This introduces a new problem, however, as the default `time-reopen()`
is 60 and not 600. We decided to let it stay 60, and make a "breaking"
change with in in `afuser`. We think, that it is not a big issue, and
as we give an option to change it back to 600, the original behavior
can be achieved with config modification.

As `afuser` does not have `logwriter_options`, and it does not
inherit from neither `logthrfetcher`, nor `logthrdest`, it has to
implement its own `time_reopen` logic.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>