kafka: fix missing log_msg_unref in case of failure
We increase the ref-count when we publish the message to the kafka broker,
and normally decrease it in the delivery callback.
However if the publishing fails, the delivery callback is never called,
though the message re-publishing will be re-tried again increasing the ref-count of the same message.

After certain number of retries an integer overflow can happen to the ref-counter,
causing an abort on the next call of log_msg_ref.

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>