msg-format: fix out-of-bounds read in msg_format_inject_parse_error()
msg_format_inject_parse_error() uses the string before that error position,
which can read from buf[-1] in case position is zero, potentially
leaking the byte in front of the data buffer, or if that address
is unmapped, can also cause a SIGSEGV.

In reality, that byte is part of the heap allocation header, so it shouldn't
be unmapped, so SIGSEGV is not very probable, at least on common platforms.

This patch also changes to using a dynamic buffer instead of a
statically sized one, avoiding the truncation of the message at 2048 bytes.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>