logmsg: clear message-type related flags in log_msg_clear()
Previously we retained flags like "mark" or "internal" which means that
these values leak through a log_msg_clear(), but if we consider
log_msg_clear() to be a function that gives us an empty slate, we should
get rid of those flags too. Same as if we created a new LogMessage instance.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>