syslog-format: fix the condition for the "kernel" workaround
The condition to use "kernel" as program name depended on
LogMessage->flags set prior to the message being parsed.

Since the location where these flags were set is pretty far from
the syslog message parser code, this dependency was not easy to
recognize and the previous set of refactoring steps even broke the
assumption.

I've decided to make the dependency clearer while retaining the workaround
where it is today:
  * LF_INTERNAL check is removed, the syslog parser is never invoked on
    internal() logs (historically it probably was...)
  * the check on LF_LOCAL was converted to a parse_options->flags check
    on LP_LOCAL. Parse_options clearly affects how the message is parsed and is
    used in a number of different locations within the same function.
  * the check on <pri> value is retained, as parsing the pri value is logically
    in the same location.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>