dbld/release: create dbld/{install,release,build} dir before docker does
dbld/rules in some cases may start docker before calling mkdir for the
above listed directories. This may lead to permission issues, as docker
runs as root and creates those directories with root, but the user
calling dbld/rules may not be root user, thus not able to utilize those
directories.

A possible solution would be to make sure in every execution path the
directories are created before docker run. That is full of trial and
error.

This alternative solution adds those directories to git, and as such
make sures the files exists before even the user able to run dbld/rules.

Signed-off-by: Kokan <kokaipeter@gmail.com>