logpipe: move cfg NULL check from caller site
We should always have cfg. If we do not, that is a coding error, so an
assert fits better for this purpose.

The motivation of doing this now is the time_reopen's inherited value.
Previously the code was prepared in a way, that when we did not have
a cfg, the time_reopen could stay -1, which could cause issues in the
timers using it. In reality cfg was always available, but this recurring
code snippet was misleading.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>