logthrfetcher: support time-reopen() on driver level
It already had a support for time reopen, and even for the possibility
to use a value other than the one inherited from the global options
(see log_threaded_fetcher_driver_init_method), but the config option
itself was not introduced to the grammar.

This grammar rule has an effect on all drivers, whose grammar utilize
`threaded_fetcher_driver_option`, such as `threaded-diskq-source` and
`python`.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>