logwriter: support time-reopen() on driver level
It already had a support for time reopen, and even for the possibility
to use a value other than the one inherited from the global options
(see log_writer_options_init), but the config option itself was not
introduced to the grammar.

This grammar rule has an effect on all drivers, whose grammar utilize
`dest_writer_option`, such as `affile`, `afprog`, `afsocket`.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>