afsocket: support time-reopen() on driver level
`afsocket` seemed to have a `time_reopen` on its own, additionally to
the `logwriter_option`'s `time_reopen` value.

An earlier commit cleaned up the usage of `time_reopen` in
`logwriter_options`, so it is better for these drivers to use that,
instead of implementing their own logic again.

In reality they could not even to that, even if they wanted, because the
`KW_TIME_REOPEN` grammar rule, inherited from `dest_writer_option` would
make it impossible to a driver, such as `afsocket`, to create their own
`KW_TIME_REOPEN` rule, because it would cause a yacc conflict.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>