github-actions: remove python2 tests
As python2 became EOL and as every syslog-ng packager switched to
python3 (except some old platforms in OBS) we no longer feel the need
to support python2.

We do not plan to remove the python2 related code yet, because it would
take a lot of effort, but we can easily remove the python2 related
tests from the CI. This means with future changes we might break
python2, and we will not see it, but we do not plan to support it
anyways.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>