logthreaded_dest_driver_init_method: add persist_name as a parameter
Continuation of #3559.

We moved the log_threaded_dest_driver_init_method in each ancesstor
to a place where it is safe to call the generate_persist_name virtual
function. But the code failed to express the reason behind it.

This PR adds the persist_name as a parameter of the public function
to make that dependency between those methods explicit.

Trying to enforce that interface, LogThreadedDestDriver stores the
provided value and asserts on it at the key points of the drivers
lifecycle.

Signed-off-by: Laszlo Szemere <laszlo.szemere@oneidentity.com>