cfg-lex: lookup already defined arguments in block argument definitions
example:
```
block destination test_http (
  body-parts("host=$HOST severity=$LEVEL")
  body("$(format-json --scope none `body-parts`)")
  ...)
{
          http(
            method("POST")
            url("http://localhost:8888")
            body(`body`)
            `__VARARGS__`
            );
};
```

Motivation:

when we have an existing SCL, like this:

```
block destination test_http (
  body-parts("host=$HOST severity=$LEVEL")
  ...)
{
          http(
            method("POST")
            url("http://localhost:8888")
            body("$(format-json --scope none `body-parts`)")
            `__VARARGS__`
            );
};
```

and `body-parts` cannot express what we really need, we have to create
a new SCL. With resolving already defined block arguments we can avoid unnecessary
copy-pastes and we can also keep backward compatibility...
On the other hand - this works only for already defined arguments. So
it won't work if the order of `body-part` and `body` are replaced.

Signed-off-by: Laszlo Budai <laszlo.budai@outlook.com>