kafka: stop keyword lookup in config() block
The keywords are dynamic in syslog-ng based on the actual context, and
in some context something should be a keyword, in other context it
shouldn't. The *retries* is a keyword defined on global level, but in
the destination/kafka/config context, that should not be a keyword.
In fact in the destination/kafka/config context nothing should be a
keyword.

```
@version: 3.30

log {
  destination {
    kafka-c(
      topic("test-topic")
      bootstrap-servers("localhost:9092")
      config(
          retries => 1
      )
    );
  };
};
```

Reported-by: Gabor Nagy <gabor.nagy@oneidentity.com>
Signed-off-by: Kokan <kokaipeter@gmail.com>