mongodb: add template support for the collection() option
The collection option now accepts templates as well.
When using literal strings, the behavior is the same as before.

In case real templates are used, the collection corresponding to the
formatted template is selected for each message.

Note that this is not a breaking change, because real templates contain
at least one $ symbol, and MongoDB collection names can not contain such a
character:

> Collection names should begin with an underscore or a letter character,
> and cannot contain the $.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>