mongodb: lower default server selection timeout to 3 seconds
By default, serverselectiontimeoutms is 30 seconds, which is unnecessarily
long time in case of syslog-ng, because we have our own reconnect mechanism
(time-reopen()).

Previously (in single-threaded mode), this option was not used at all,
because "serverSelectionTryOnce" was active. ("serverSelectionTryOnce" is
not available when pooled clients are used.)

This commit sets the default server selection timeout to 3 seconds, which
avoids syslog-ng waiting too long to reload/shutdown.

Unfortunately, mongoc_uri_set_option_as_int32() is not available in
really old mongoc versions (<1.4), hence the conditional build flag.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>