java: set log4j log level based on syslog-ng's internal log level
This commit fixes Java debug logging.
Previously the log4j log level was left on its default value, even when
syslog-ng was started in debug/trace mode.

Using an enum would be cleaner, but InternalMessageSender uses integers
instead as they are easier to handle on the JNI side.

Note: Currently, the log level is not updated when debug/trace is enabled
or disabled at runtime (using syslog-ng-ctl).

Signed-off-by: László Várady <laszlo.varady@protonmail.com>