driver: rename acquire_queue_method() to acquire_memory_queue()
This method is responsible only for creating memory queues.
Other type of queues can be created by overriding the acquire_queue()
virtual method of the LogDestDriver.

This is a preparation step for fixing a crash caused by a queue type
mismatch (when cfg_persist_config_fetch() returns a disk-based queue).

Signed-off-by: László Várady <laszlo.varady@protonmail.com>