driver: fix crash when switching between disk-based and memory queues
When a disk-buffer is removed from the configuration, and syslog-ng is
reloaded, syslog-ng starts acquiring a memory queue instead, but first it
fetches and tries to use the previous queue from cfg-persist. The type
of the old queue is not validated, so we end up using a different type of
queue than what the user configured.

This is not a disaster, this behavior could have been documented, but
we also crash when trying to use the old queue, because it was not
initialized with the queue's own initialization method
(DiskQ::acquire_queue()).

This commit fixes the crash by safely closing the disk-buffer queue,
and allocating a memory queue instead (exactly what the user wants; and
unsent messages remain available in the disk buffer file).

Signed-off-by: László Várady <laszlo.varady@protonmail.com>