kafka: batch for sync_send(yes)
With sync_send(yes), batching actually still calls produce to librdkafka (so messages are delivered), but the transaction commit is only done per batching size. The transaction commit is the blocking operation.

With sync_send(no), batching parameters can be configured. But that has no particular effect. It would be possible to only call produce per batch size, but the produce itself is not that expensive (compared to sending messages to kafka) and a vector should be manually updated.

Signed-off-by: Kokan <kokaipeter@gmail.com>