kafka: make kafka options protected from overwrites
Some options of kafka can be overriden from the config() and properties-file(), e.g. bootstrap-servers().
We consider this as a potential misconfiguration, so we don't allow setting these properties from config() or properties-file().

The only prohibited property is "bootstrap-servers" (alias "metadata.broker.list"), which would override bootstrap-servers() option of kafka driver.
Topic(), or key() options are not affected (there is no librdkafka property for those).

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>