logmpx: fix write protecting LogMessage
In LogMultiplexer, checking the number of next_hops is not enough to
determine the number of LogPipes, that will get the current LogMessage,
because we need to take account of pipe_next, too.

See the following log path, for example:
```
log {
  source(s_network);
  destination(d_http);
  rewrite(r_set)
};
```
Here, the LogMultiplexer has one next_hop (d_http) and one pipe_next (r_set).
In this case we do not set the write protection, so the r_set rewrite
rule will modify the same LogMessage that is queued for d_http.
If the LogMessage gets handled in the threaded destination (d_http)
after r_set has changed it, we get an unexpected behavior, in a bad case
even a crash.

========================================================================

The other problem is that we remove the write protection before
forwarding the LogMessage to pipe_next.

In the example above we set write protection, queue the message to
d_http, remove the write protection, then modify the same message in
r_set. When d_http sends the message a bit later, the message is already
modified.

We should not remove the write protection in LogMultiplexer, so we can
delete the write protection removal code from there.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>