kafka: check remaining messages after the purge only in main thread
Before this PR, only the main thread initiated flush and purge, so warning the user if there are remaining messages after the purge was okay.
However this is changed to workers be able to shut down even from blocking produce calls, so we shouldn't warn the user for remaining message after a purge is called from the worker thread.

The reason is that more purge calls will be called from other worker threads and from the main thread too.

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>