qdisk: use gint64 instead of off_t in _truncate_file()
`off_t` and `gint64` are compatible both ways on 64 bit machines.
We were relying on this implicitly before, too.

It is enough to use `off_t` exactly where the system calls are made, our
private functions should use glib types.

With this commit we can clearly see the lines where the casting happens,
and we only use `off_t` where we are near system calls.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>