qdisk: don't truncate if it would make a small size difference
ftruncate() is an expensive call.

In the current form, in case of a reliable disk-buffer when the
destination acks the message, the truncate is usually called.

Let's look at the situation, where the destination can keep up with the
speed of the source:
  1. Incoming message, it gets written to the disk buffer, the write
     head moves to RESERVED_SPACE + msg_size.
  2. We read the message from the disk-buffer, and send it to the
     destination, the read head moves to RESERVED_SPACE + msg_size.
  3. The destination acks the message, the backlog_head moves to
     RESERVED_SPACE + msg_size.
  4. At the end of the ack, we see, that the disk buffer is "empty",
     meaning all heads are at the same position, so we truncate the file
     to RESERVED_SPACE, and move all 3 heads there.
  5. A new message arrives, and we go to step 1.

In this scenario we call ftruncate() for every incoming message.
I have made a test code, which measures the fwrite, fread, ftruncate
iteration's time taken, and I observed, that more than half of the
real word time was taken by ftruncate():
```
Writing, reading 1048576 * 1024 bytes messages, truncating the file to 4096 bytes per message.
  [1/10] DONE in 7.856170541 seconds, from which ftruncate() took 4.734487726 seconds.
  [2/10] DONE in 7.773038602 seconds, from which ftruncate() took 4.678402817 seconds.
  [3/10] DONE in 7.780979727 seconds, from which ftruncate() took 4.664749331 seconds.
  [4/10] DONE in 7.805741840 seconds, from which ftruncate() took 4.675652459 seconds.
  [5/10] DONE in 7.745974862 seconds, from which ftruncate() took 4.642292874 seconds.
  [6/10] DONE in 7.800890915 seconds, from which ftruncate() took 4.690607455 seconds.
  [7/10] DONE in 7.774325797 seconds, from which ftruncate() took 4.693368755 seconds.
  [8/10] DONE in 7.806343633 seconds, from which ftruncate() took 4.683172069 seconds.
  [9/10] DONE in 7.896176455 seconds, from which ftruncate() took 4.757315806 seconds.
  [10/10] DONE in 7.868837540 seconds, from which ftruncate() took 4.715273111 seconds.
```
I've done the measurements on an SSD.

To optimize the calls of ftruncate(), in this commit I have introduced a
new logic: we calculate the possible size reduction in each
_truncate_file() call, and we do not call ftruncate() if it is less,
than 1 percent of the maximal disk-buffer size.

My measurements changed drastically when implementing the same logic in
the test code:
```
Writing, reading 1048576 * 1024 bytes messages, truncating the file to 4096 bytes if the size reduction reaches 1% of max size.
  [1/10] DONE in 2.255469215 seconds, from which ftruncate() took 0.107140776 seconds.
  [2/10] DONE in 2.226006879 seconds, from which ftruncate() took 0.109389325 seconds.
  [3/10] DONE in 2.239258063 seconds, from which ftruncate() took 0.110315009 seconds.
  [4/10] DONE in 2.229625672 seconds, from which ftruncate() took 0.110243967 seconds.
  [5/10] DONE in 2.255923025 seconds, from which ftruncate() took 0.108881297 seconds.
  [6/10] DONE in 2.223326513 seconds, from which ftruncate() took 0.107711562 seconds.
  [7/10] DONE in 2.251854651 seconds, from which ftruncate() took 0.134951278 seconds.
  [8/10] DONE in 2.235311433 seconds, from which ftruncate() took 0.126703875 seconds.
  [9/10] DONE in 2.228932757 seconds, from which ftruncate() took 0.109558716 seconds.
  [10/10] DONE in 2.259198891 seconds, from which ftruncate() took 0.109496328 seconds.
```

This logic can cause the disk-buffer file to be maximum 1% in size of
the maximal size, with having only already processed messages in it.
I believe it is not a big tradeoff, when seeing how much time it can
save.

I have done measurements with other thresholds:
 1. 50% of max size: took the same time
 2. 0.1%, 0.01%, 0.001%: took nearly the same time
 3. 0.0001%, 0.00001%: took as much as read and write

What I observed is that it gets slow, when truncating nearly per msg.
Then we reach a point (approximately at every 500th message), where no
matter if we truncate at 1% or 50%, it takes the same time.

Thinking about a general config, where the disk buffer is 1GB, its 1% is
10MB, so if we receive smaller then 20KB sized messages, we have the
optimal truncate time. Or for a 100MB disk-buffer it is 2KB.

Max 2KB message size and 100MB disk-buffer are average values, so I
think the 1% of the max size threshold works most of the time.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>