logqueue-disk: move write_message() to the descendant classes
This functions does 3 things:
 1. Does some preliminary check on QDisk.
 2. Serializes the message.
 3. Writes the message to the QDisk.

Serializing and writing to disk are expensive actions.
We are doing both, while holding the LogQueue::lock.
We do not need to hold the lock while serializing.

In order to move the serialization outside the lock, we need more
control, which cannot be done with this common function.

This commit only moves the function to the descendant classes, we will
implement our class specific behavior in the following commits.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>