diskq: fix memory leak when force-suspending source
log_queue_disk_drop_message() is responsible for dropping a message
when the disk-buffer reaches its full capacity.

In case flow-control is enabled, the source will stop before the queue
becomes full, so no message loss is possible, but there are special/rare
cases when the user configures mem-buf-length() or mem-buf-size()
incorrectly, and the message has to be dropped.

log_queue_disk_drop_message() must force-suspend the source in the above
case, but we had already lost one single message by then.

log_msg_drop() not only acks the message with the specified ack type,
it also calls log_msg_unref() on the message, which should be done in
both cases to avoid leaking the memory.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>